requirements.txt
black==23.9.1
cfgv==3.4.0
click==8.1.7
distlib==0.3.7
filelock==3.12.4
flake8==6.1.0
flake8-quotes==3.3.2
identify==2.5.30
isort==5.12.0
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pre-commit==3.4.0
pyaes==1.6.1
pyasn1==0.5.0
pycodestyle==2.11.0
pyflakes==3.1.0
python-dotenv==1.0.0
PyYAML==6.0.1
rsa==4.9
Telethon==1.30.3
tomli==2.0.1
typing_extensions==4.8.0
virtualenv==20.24.5
